Restructured DQM Bokeh app, in order to have working app AND working unit tests
